+++ muos/steps/end.py
@@ -2,7 +2,7 @@ from ..environment import Environment
 from ..step import Step
 
 class End(Step):
-    description: str = 'End'
+    description: str = 'Ending of the runner...'
 
     def main(self, environment: Environment) -> None:
         super().main(environment)