+++ muos/steps/unmount_partitions.py
@@ -3,7 +3,7 @@ from ..environment import Environment
 from ..step import Step
 
 class UnmountPartitions(Step):
-    description: str = 'Unmount partitions'
+    description: str = 'Unmounting partitions...'
 
     def main(self, environment: Environment) -> None:
         super().main(environment)