+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Improve `Log.message()` from `Runner.run()`.
 - Improve timing of calling `end` step.
 - Rename `Runner.name` to `Runner.description`.
+- Improve descriptions in steps.
 
 ## [22.09.0a1] - 2022-09-10
 ### Added